bootstrap: split controller-runtime users out of nodeup-linked packages
Move the PKI bootstrap verifier to pkg/bootstrap/pkibootstrap/pkiverifier
and the Cluster API machine lookup to
pkg/nodeidentity/clusterapi/capimanager. Both are only used by
kops-controller, but they lived in packages that nodeup links for the
authenticator and the Machine type, and their controller-runtime client
import registers the full Kubernetes API scheme, which dead-code
elimination cannot remove.
